Fix cluster/clean target & install script
Locally in minishift I've found that when the AMO starts & reacts to
the CR, errors are thrown about the prometheus kinds not being
found. This appears to be because, well, they didn't exist when it
started (they're created by the prometheus operator?), and the AMO
can't find them in its cache probably.

I also noticed errors being ignored about the role/rolebinding being
done in the wrong order, but execution continues anyway, so I've just
explicitly put them in the right order, and also put in a simple check
to make sure that the SA has the right access.

I also updated the cluster/clean target since that was broken in a
couple of ways before (causing the namespace to be undeletable without
figuring out what in there had a finalizer and cleaning it up). Also
the cluster/prepare was referencing a file that didn't exist.